requirements.txt
asgiref==3.11.1
brotli==1.2.0
click==8.3.1
colorama==0.4.6
distlib==0.4.0
dj-database-url==3.1.0
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
filelock==3.21.2
gunicorn==21.2.0
h11==0.16.0
inflection==0.5.1
packaging==26.0
platformdirs==4.7.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
uvicorn==0.40.0
virtualenv==20.36.1
whitenoise==6.6.0
